Changed axis to not display negative values + overflow bug fixes
- Added relative API to Range object to allow the AxisComponent to
  work on relative timestamps. This fixed an issue where the tick
  markers were not zero'ed at the profiling start time
- Fixed an interger overflow introduced when changing time unit from
  ms to us, which caused the time axis to not render the labels properly

+++ adt-ui/src/test/java/com/android/tools/adtui/AxisFormatterTest.java
@@ -38,7 +38,7 @@ public class AxisFormatterTest extends TestCase {
     // maxMinorTicks is 5 so the smallest possible interval is 100 / 5 = 20
     // The axis is in base 10 which has factors: {1, 5, 10}, so the interval
     // will get round up using the factor 5, which gives us 50.
-    int interval = domain.getMinorInterval(100);
+    long interval = domain.getMinorInterval(100);
     assertEquals(50, interval);
 
     // maxMajorTicks is 10 so the smallest possible interval is 1000 / 10 = 100
